Moving field type inference code to generic implementation. This is for code reviewing purposes only, as there may be a memory leak bug in this implementation.
